[30m] emprolijo la estructura de archivsos
