+++ public/js6/main.js
@@ -1,13 +1,6 @@
-import React from 'react';
 import ReactDOM from 'react-dom';
-import "whatwg-fetch";
+import React from 'react';
 import {App} from "./App";
-
-import {Login} from "./Login";
-import {MainScreen} from "./MainScreen";
-import {BusinessUsers} from "./BusinessUsers";
-import {Servers} from "./Servers";
-import {Users} from "./Users";
