+++ public/js6/table/CreateDialog.js
