+++ public/js6/table/Dialog.js
