+++ public/js6/table/FilterDialog.js
