+++ public/js6/table/PagingDialog.js
