+++ public/js6/tables/rules/ViewCommitsButton.js
