+++ public/js6/tables/servers/TokenCreatorButton.js
