+++ public/js6/tables/users/Cars.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import "whatwg-fetch";
 import Popout from 'react-popout';
-import {CrudTable} from "./CrudTable"
+import {CrudTable} from "../../table/CrudTable"
 
 class Strategy{
     constructor(token,userId){
